prepare CRUDs for unit testing

+++ src/users/services/users.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnprocessableEntityException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm/dist';
 import { QueryRunner, Repository } from 'typeorm';
 import { User } from '../entities';
 import { CreateUserDTO } from '../dtos';
 import { Authentication } from 'src/auth/entities';
+import { TypeORMError } from 'typeorm';
 
 @Injectable()
 export class UsersService {
@@ -17,15 +18,32 @@ export class UsersService {
   }
 
   async findOne(id: number): Promise<User | null> {
+    if (typeof id !== 'number') {
+      throw new UnprocessableEntityException('id must be of type number');
+    }
     return this._usersRepository.findOneBy({ id });
   }
 
   async remove(id: number): Promise<void> {
-    await this._usersRepository.delete(id);
+    const user = await this.findOne((id = id));
+    if (typeof id !== 'number') {
+      throw new UnprocessableEntityException('id must be of type number');
+    } else if (user !== null) {
+      await this._usersRepository.delete(id);
+    } else {
+      throw new NotFoundException('there is no user with that id');
+    }
   }
 
-  async create(createUserDTO: CreateUserDTO, authentication: Authentication, queryRunner: QueryRunner): Promise<User> {
-    const user = this._usersRepository.create({ ...createUserDTO, authentication });
-    return queryRunner.manager.save(user);
+  async create(createUserDTO: CreateUserDTO, authentication: Authentication | undefined, queryRunner: QueryRunner): Promise<User | undefined> {
+    let user: User | undefined = undefined;
+    try {
+      user = this._usersRepository.create({ ...createUserDTO, authentication });
+      return queryRunner.manager.save(user);
+    } catch (e) {
+      throw new TypeORMError('something went wrong with provided dto or authentication entity');
+    } finally {
+      return user;
+    }
   }
 }
